production_system/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.1.3
Flask==3.0.0
Flask-RESTful==0.3.10
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.26.1
packaging==23.2
pandas==2.1.2
pluggy==1.3.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
threadpoolctl==3.2.0
tomli==2.0.1
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
